VR-175: moving files part #2, combining scope3 views.
