+++ src/controllers/queries/__tests__/helpers.ts
@@ -7,10 +7,10 @@ import Database from '../../../models/db/index.js'
 import { ConnectionRow, QueryRow } from '../../../models/db/types.js'
 import { UUID } from '../../../models/strings.js'
 import VeritableCloudagent from '../../../models/veritableCloudagent.js'
-import Scope3CarbonConsumptionResponseTemplates, { Scope3FormProps } from '../../../views/queries/co2scope3.js'
-import Scope3CarbonConsumptionTemplates from '../../../views/queries/newCo2scope3.js'
 import QueriesTemplates from '../../../views/queries/queries.js'
 import QueryListTemplates from '../../../views/queries/queriesList.js'
+import Scope3CarbonConsumptionTemplates from '../../../views/queries/requestCo2scope3.js'
+import Scope3CarbonConsumptionResponseTemplates, { Scope3FormProps } from '../../../views/queries/responseCo2scope3.js'
 
 type QueryStatus = 'resolved' | 'pending_your_input' | 'pending_their_input'
 
@@ -76,7 +76,7 @@ const defaultOptions: QueryMockOptions = {
   },
 }
 
-function templateFake(templateName: string, props?: any) {
+function templateFake(templateName: string, props?: Scope3FormProps) {
   if (props?.partial) return Promise.resolve(`${templateName}_template-${JSON.stringify(props)}`)
   return Promise.resolve(`${templateName}_template`)
 }
@@ -111,7 +111,7 @@ export const withQueriesMocks = (testOptions: Partial<QueryMockOptions> = {}) =>
   } as unknown as QueryListTemplates
   const mockLogger: ILogger = pino({ level: 'silent' })
   const dbMock = {
-    get: sinon.stub().callsFake((tableName: 'connection' | 'query', {}) => Promise.resolve(options.getRows[tableName])),
+    get: sinon.stub().callsFake((tableName: 'connection' | 'query', _) => Promise.resolve(options.getRows[tableName])),
     update: sinon.stub().resolves(),
     insert: sinon.stub().resolves([
       {
